Move primary client stuff into a client factory.
This allows tests to be done more easily, since now the client can be
loaded without starting it.
`index.js` now just pushes the config to the client factory and starts
the resulting client.
